#1 refactor : 공통 컴포넌트 파일 경로 변경
