+++ src/components/common/ActivityCard.tsx
