+++ src/components/common/ContentCard.tsx
