+++ src/components/common/StatCard.tsx
